Suppann (#53)
* Ajout de la page produit à developement (#22)

* Profile page (#16)

* page profil finie + responsive

* rajout du bouton création

* style: run linter

* fix: structure

* avancement de la page detail du produit

* avancement de la page produit

* Page de produit finie

* petits changements sur page produit

* page register et login (#26)

* page register et login simple et efficace finie

* changement du css

* déplacement des erreurs sous les inputs

* feat: admin graphs (2) (#25)

- Accounts/bans over time
- Accounts created by day

* page index (#23)

* fix connect db pour windows =)

* filtres index style et responsive

* style moche

* style: run linter

---------


* ui: dash bans table (#27)

- Latest bans table
- Search bans
- Unban system

* feat: logins by day graph

* Val (#30)

* fix: details admin btns (#31)

* Posts model & ice (#32)

* fix: user not defined sur page register

* feat: add admin route to header and header to admin page

* feat: add more admin graphs (#36)

* feat: add posts db populate helper (#37)

* refactor: dashboard responsive and style (#38)

* refactor: uniformisation du style (#39)

* post system (#34)

* ajout post a la db temporaire

* ajout des annonces dans la db + affichage sur index

* refactoring posts code

* affichage des nouvelles annonces en premier sur index

* ajout min max prix annonce

* fix crash avec /profile

* fix index page UI bug

* ajout fonction getOnePost avec un id de post

* fix: rework database functions and model

* feat: add errors on post creation

* fix: main page post display

* feat: post images upload

* style: run linter

---------


* refactor: suppression de la pp sur la page profil

* fix: remove test

* fix UI bug <a> index

* placeholder form annonce + fix padding

* lien entre index et details + annonces sur page profil (#42)

* lien entre la page index et la page details

* rajout des annonces sur page profil

* annonces clickable dans profile

* passage de EN à FR

* faute d'ortho rectifié

* fix: affichage des images des annonces sur page profil

* rajout d'une simple page d'erreur 404 (#46)

* Ban user (#47)

* ajout du ban user depuis une annonce

* ajout commentaires et espaces dans le code

* pareil

* bouton supprimer annonce

* ajout message confirmation avant suppression de l'annonce

* style: run linter

* retirer description annonces page index

---------


* avancement de suppression annonce

* refactor: change api endpoints

* feat: post suppression

* fixes

* fix: delete posts when user is banned

* style: run linter

---------

+++ src/routes/post.js
@@ -9,12 +9,12 @@ const {
   setPictures,
   getPost,
   deletePost,
+  getPostById,
 } = require("../database/post");
+const { hasAlreadyGivenRank, addRank } = require("../database/ranks");
+const { isUserPostOwner } = require("../utils/utils");
 
-const { banUser, getUserByEmail } = require("../database/auth");
-const { addRank, hasAlreadyGivenRank } = require("../database/ranks");
-
-router.get("/profile/create", authMiddleware.userAuth, (req, res) => {
+router.get("/create", authMiddleware.userAuth, (req, res) => {
   const errors = req.session.errors;
   const formData = req.session.formData;
   const user = req.session.user;
@@ -33,7 +33,7 @@ const createSchema = z.object({
   image: z.array(z.string()).optional(),
 });
 
-router.post("/profile/create", authMiddleware.userAuth, async (req, res) => {
+router.post("/create", authMiddleware.userAuth, async (req, res) => {
   const body = req.body;
 
   body.price = parseInt(body.price);
@@ -52,7 +52,7 @@ router.post("/profile/create", authMiddleware.userAuth, async (req, res) => {
     req.session.errors = errors;
     req.session.formData = body;
 
-    return res.redirect("/profile/create");
+    return res.redirect("/posts/create");
   }
 
   const userID = req.session.user.id;
@@ -75,28 +75,35 @@ router.post("/profile/create", authMiddleware.userAuth, async (req, res) => {
 });
 
 // Detailed post page route
-router.get("/post/:id", async (req, res) => {
+
+router.get("/:id", async (req, res) => {
   const id = req.params.id;
   const post = await getPost(id);
 
   return res.render("pages/details", { post, user: req.session.user });
 });
 
-// Handles the ban of a user with an admin account
-router.post("/banUser/:email", authMiddleware.adminAuth, async (req, res) => {
-  const { reasonForBan } = req.body;
-  const userEmail = req.params.email;
-  await banUser(userEmail, reasonForBan);
-  res.redirect("/");
-});
-
-router.post("/deletePost/:id", authMiddleware.adminAuth, async (req, res) => {
+router.post("/delete/:id", authMiddleware.userAuth, async (req, res) => {
   const postID = req.params.id;
+
+  const post = await getPostById(postID);
+
+  if (!post) {
+    console.log("post not found");
+    return res.redirect("/");
+  }
+
+  // post owner / admin
+  if (!isUserPostOwner(req.session.user, post)) {
+    console.log("not the owner");
+    return res.redirect("/");
+  }
+
   await deletePost(postID);
   res.redirect("/");
 });
 
-router.post("/giveRank/:id", authMiddleware.userAuth, async (req, res) => {
+router.post("/addrank/:id", authMiddleware.userAuth, async (req, res) => {
   const postID = req.params.id;
   const { givenRank } = req.body;
   const userID = req.session.user.id;
